+++ examples/docs_snippets/docs_snippets/guides/external-systems/assets-with-io-managers.py
@@ -21,17 +21,3 @@ def clean_sales_data(raw_sales_data: pd.DataFrame) -> pd.DataFrame:
 @dg.asset
 def sales_summary(clean_sales_data: pd.DataFrame) -> pd.DataFrame:
     return clean_sales_data.groupby(["owner"])["amount"].sum().reset_index()
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        # highlight-start
-        # Define the I/O manager and pass it to `Definitions`
-        resources={
-            "io_manager": DuckDBPandasIOManager(
-                database="sales.duckdb", schema="public"
-            )
-        }
-        # highlight-end
-    )
