+++ examples/docs_snippets/docs_snippets/guides/external-systems/assets-with-snowflake-io-manager.py
@@ -1,5 +1,4 @@
 import pandas as pd
-from dagster_snowflake_pandas import SnowflakePandasIOManager
 
 import dagster as dg
 
@@ -17,20 +16,3 @@ def clean_sales_data(raw_sales_data: pd.DataFrame) -> pd.DataFrame:
 @dg.asset
 def sales_summary(clean_sales_data: pd.DataFrame) -> pd.DataFrame:
     return clean_sales_data.groupby(["owner"])["amount"].sum().reset_index()
-
-
-@dg.definitions
-def resources():
-    return dg.Definitions(
-        resources={
-            # highlight-start
-            # Swap in a Snowflake I/O manager
-            "io_manager": SnowflakePandasIOManager(
-                database=dg.EnvVar("SNOWFLAKE_DATABASE"),
-                account=dg.EnvVar("SNOWFLAKE_ACCOUNT"),
-                user=dg.EnvVar("SNOWFLAKE_USER"),
-                password=dg.EnvVar("SNOWFLAKE_PASSWORD"),
-            )
-            # highlight-end
-        }
-    )
