+++ examples/docs_snippets/docs_snippets/guides/external-systems/resources-without-io-managers.py
@@ -1,0 +1,10 @@
+from dagster_duckdb_pandas import DuckDBResource
+
+import dagster as dg
+
+
+@dg.definitions
+def resources():
+    return dg.Definitions(
+        resources={"duckdb": DuckDBResource(database="sales.duckdb", schema="public")}
+    )
